[FD] Don't start up a TestServer in tests that don't need it
Tests like DesAgentClientApiConnectorISpec don't need to make HTTP requests to a running version of the application, so don't indirectly inherit OneServerPerSuite to avoid starting up an unnecessary TestServer.

Also push the implicit HeaderCarrier down to the tests that need it to avoid surprises and make wireMockAppConfiguration easier to read by inlining only-used-once constants.